+++ MAREngine/src/Core/graphics/Mesh/Vertex.h
@@ -32,7 +32,7 @@ namespace marengine {
 		maths::vec3 position;
 		maths::vec3 lightNormal;
 		maths::vec2 textureCoordinates;
-		float shapeID{ 0.f };
+		uint32_t shapeID{ 0 };
 	};
